+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## 1.0.2 under development
 
-- Bug #32: On throw exception in process getting from container optional dependency use default value only if container
-  don't have dependency or thrown not `CircularReferenceException` (vjik) 
+- Bug #32: Throw exception instead of returning default value if optional dependency exists but there is an exception when getting it (vjik) 
 
 ## 1.0.1 December 19, 2021