+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.2 under development
 
-- no changes in this release.
+- Bug #32: Throw exception instead of returning default value if optional dependency exists but there is an exception when getting it (vjik) 
 
 ## 1.0.1 December 19, 2021